feat: 역 등록시 중복 이름 검증 기능 구현 fix: 노선 수정시, 이름을 변경하지 않는 경우, 노선명 중복 예외 발생 버그 수정 feat: 존재하지 않는 노선에 대해 삭제 요청하는 경우, 예외 발생 구현 test: 테스트 케이스 추가 구현
StationController 역 추가 기능 중 중복 역명 검증 할수 있도록 로직 추가
StationRepository 역명을 통한 중복 검증 위한 검증 메서드 추가
Line 역명 중복 검증 로직에 필요한 두 객체의 이름 비교 기능 구현
LineService 노선 수정시 기존의 이름 그대로 사용하는 경우 발생했던 버그 수정
LineService 존재하지 않는 노선 삭제시도시 예외 발생 구현

+++ src/main/java/wooteco/subway/admin/controller/StationController.java
@@ -30,7 +30,11 @@ public class StationController {
 	@PostMapping
 	public ResponseEntity<StationResponse> createStation(@RequestBody Request<StationCreateRequest> view) {
 		StationCreateRequest content = view.getContent();
-		Station persistStation = stationRepository.save(content.toStation());
+		Station entity = content.toStation();
+		if (stationRepository.existsByName(entity.getName())) {
+			throw new IllegalArgumentException("이미 저장되어있는 역이름입니다.");
+		}
+		Station persistStation = stationRepository.save(entity);
 
 		return ResponseEntity
 			.created(URI.create("/stations/" + persistStation.getId()))
@@ -44,6 +48,9 @@ public class StationController {
 
 	@DeleteMapping("/{id}")
 	public ResponseEntity<Void> deleteStation(@PathVariable Long id) {
+		if (!stationRepository.existsById(id)) {
+			throw new IllegalArgumentException("존재하지 않는 리소스는 이용할 수 없습니다.");
+		}
 		stationRepository.deleteById(id);
 		return ResponseEntity.noContent().build();
 	}
